AIR-QUALITY/requirements.txt
geopy
python-dotenv
pytz
requests
streamlit
plyer